backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
ecdsa==0.19.1
fastapi==0.115.13
greenlet==3.2.3
h11==0.16.0
idna==3.10
pip>=23.3
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic[email]==2.11.7
pydantic_core==2.33.2
python-jose==3.5.0
python-multipart>=0.0.18
rsa==4.9.1
setuptools>=78.1.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
python-dotenv==1.0.0
